protect endpoints

+++ core/views.py
@@ -1,16 +1,14 @@
 from rest_framework.decorators import api_view, permission_classes
-from rest_framework.permissions import IsAuthenticated, IsAuthenticatedOrReadOnly
 from rest_framework.response import Response
+from rest_framework.permissions import AllowAny
 from rest_framework import status
-from accounts import models as accounts_models
 from utils import main
 from . import models
 from . import serializers
-from . import permissions
 
 
 @api_view(["GET", "POST"])
-@permission_classes([IsAuthenticatedOrReadOnly])
+@permission_classes([AllowAny])
 def innovation_list(request):
     """
     List all innovations, or create a new innovation
@@ -21,6 +19,11 @@ def innovation_list(request):
         return paginated_response
 
     elif request.method == "POST":
+        if not request.user.is_authenticated:
+            return Response(
+                {"detail": "Authentication credentials were not provided."},
+                status=status.HTTP_401_UNAUTHORIZED,
+            )
         serializer = serializers.Innovation(
             data=request.data, context={"request": request}
         )
@@ -33,7 +36,7 @@ def innovation_list(request):
 
 
 @api_view(["GET", "PUT", "PATCH", "DELETE"])
-@permission_classes([permissions.IsInnovationAuthorOrModOrReadOnly])
+@permission_classes([AllowAny])
 def innovation_detail(request, pk):
     """
     Contains methods for updating an innovation (either partially or entirely) and deleting an innovation
@@ -47,7 +50,23 @@ def innovation_detail(request, pk):
         serializer = serializers.Innovation(innovation, context={"request": request})
         return Response(serializer.data)
 
-    elif request.method == "PUT":
+    if not request.user.is_authenticated:
+        return Response(
+            {"detail": "Authentication credentials were not provided."},
+            status=status.HTTP_401_UNAUTHORIZED,
+        )
+
+    if (
+        request.user.user_profile != innovation.author
+        or not request.user.is_site_mod
+        or not request.user.is_staff
+    ):
+        return Response(
+            {"detail": "You do not have permission to access this resource."},
+            status=status.HTTP_403_FORBIDDEN,
+        )
+
+    if request.method == "PUT":
         serializer = serializers.Innovation(
             innovation, data=request.data, context={"request": request}
         )
@@ -72,20 +91,58 @@ def innovation_detail(request, pk):
         return Response(status=status.HTTP_204_NO_CONTENT)
 
 
+@api_view(["POST"])
+def like_innovation(request, pk):
+    try:
+        innovation = models.Innovation.objects.get(pk=pk)
+    except models.Innovation.DoesNotExist:
+        return Response(status=status.HTTP_404_NOT_FOUND)
+
+    request.data["author"] = request.user.user_profile
+    request.data["innovation"] = models.Innovation.objects.get(pk=pk).pk
+    print(request.data)
+    # return Response({})
+    serializer = serializers.Like(data=request.data, context={"request": request})
+
+    if serializer.is_valid():
+        serializer.validated_data["author"] = request.user.user_profile
+        serializer.save()
+        return Response({serializer.data})
+    return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+
+def unlike_innovation(request):
+    pass
+
+
+def bookmark_innovation(request):
+    pass
+
+
+@api_view(["DELETE"])
+def unbookmark_innovation(request):
+    pass
+
+
 @api_view(["GET", "POST"])
-@permission_classes([IsAuthenticatedOrReadOnly])
+@permission_classes([AllowAny])
 def innovation_comment_list(request, pk):
     """
     List all comments for a specific innovation, or create a comment for an innovation
     """
     if request.method == "GET":
         innovation_comments = models.InnovationComment.objects.filter(innovation_id=pk)
-        serializer = serializers.InnovationComment(
-            innovation_comments, many=True, context={"request": request}
+        paginated_response = main.paginate(
+            request, innovation_comments, serializers.InnovationComment
         )
-        return Response(serializer.data)
+        return paginated_response
 
     elif request.method == "POST":
+        if not request.user.is_authenticated:
+            return Response(
+                {"detail": "Authentication credentials were not provided."},
+                status=status.HTTP_401_UNAUTHORIZED,
+            )
         request.data["innovation"] = models.Innovation.objects.get(id=pk).pk
 
         serializer = serializers.InnovationComment(
@@ -100,15 +157,13 @@ def innovation_comment_list(request, pk):
 
 
 @api_view(["GET", "PUT", "PATCH", "DELETE"])
-@permission_classes(
-[permissions.IsInnovationAuthorOrCommentAuthorOrModOrStaffOrReadOnly]
-)
-def innovation_comment_detail(request, pk):
+@permission_classes([AllowAny])
+def innovation_comment_detail(request, pk, cpk):
     """
     Contains methods for updating an innovation comment (either partially or entirely) and deleting an innovation comment
     """
     try:
-        innovation_comment = models.InnovationComment.objects.get(pk=pk)
+        innovation_comment = models.InnovationComment.objects.get(pk=cpk)
     except models.InnovationComment.DoesNotExist:
         return Response(
             {"detail": "Comment does not exist"}, status=status.HTTP_404_NOT_FOUND
@@ -121,6 +176,16 @@ def innovation_comment_detail(request, pk):
         return Response(serializer.data)
 
     elif request.method in ["PUT", "PATCH"]:
+        if (
+            request.user.user_profile != innovation_comment.author
+            or not request.user.is_site_mod
+            or not request.user.is_staff
+        ):
+            return Response(
+                {"detail": "You do not have permission to access this resource."},
+                status=status.HTTP_403_FORBIDDEN,
+            )
+
         serializer = serializers.InnovationComment(
             innovation_comment,
             data=request.data,
@@ -135,5 +200,18 @@ def innovation_comment_detail(request, pk):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     elif request.method == "DELETE":
+        if (
+            request.user.user_profile != innovation_comment.author
+            or not request.user.is_site_mod
+            or not request.user.is_staff
+        ):
+            return Response(
+                {"detail": "You do not have permission to access this resource."},
+                status=status.HTTP_403_FORBIDDEN,
+            )
         innovation_comment.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
+
+
+def bookmarks(request):
+    pass
